user auth added

+++ main/client/features/listOfStudents/studentListCtrl.js
@@ -1,25 +1,19 @@
 angular.module('reportGenerator').controller('studentListCtrl', function($scope, $element, classService) {
   var self = this;
   this.addingStudent = false;
-
   this.showCompileStuff = false;
 
   this.doReports = function(studentList) {
     self.showStudent = false;
+    var userId = self.user._id;
 
-    classService.compileReports(studentList).then(function(response, err){
-      console.log(response, "reponse from server")
-      self.reports = response.data;
-      console.log(self.reports, "reports after server")
-          self.showReports = true;
-    })
-    // self.showReports = true;
-    console.log(self.reports, "the waiting")
-
+    classService.compileReports(userId, studentList).then(function(response, err) {
+        self.reports = response.data;
+        self.showReports = true;
+      })
   };
   this.editStudent = function(student) {
     self.showStudent = true;
-    // console.log(self.studentList)
     student.classId = self.studentList._id;
     student.className = self.studentList.className;
     self.currentStudent = student;
@@ -28,43 +22,38 @@ angular.module('reportGenerator').controller('studentListCtrl', function($scope,
   };
 
   this.deleteStudent = function(studentClassId, studentId) {
-    classService.deleteStudent(studentClassId, studentId).then(function(response, err) {
+    var userId = self.user._id;
+    classService.deleteStudent(userId, studentClassId, studentId).then(function(response, err) {
 
     })
 
     for (var i = 0; i < self.studentList.students.length; i++) {
       if (self.studentList.students[i]._id == studentId) {
-      self.studentList.students.splice(i, 1);
+        self.studentList.students.splice(i, 1);
 
       }
     }
-self.classInfo = self.studentList;
-if(self.studentList.students.length == 0){
-  self.showCompileStuff= false;
-}
+    self.classInfo = self.studentList;
+    if (self.studentList.students.length == 0) {
+      self.showCompileStuff = false;
+    }
     self.showReports = false;
     self.showStudent = false;
 
   };
 
   $scope.$watch('is.classInfo', function(newValue, oldValue) {
-
     self.studentList = newValue;
     this.showCompileStuff = false;
-
   });
 
 
   $scope.$watch('is.studentList.students', function(newValue, oldValue) {
-
   });
-
 
   this.doIt = function() {
     $('select').material_select();
   }
-
-
 
   this.startAddNewStudent = function() {
     self.addingStudent = !self.addingStudent;
@@ -79,11 +68,13 @@ if(self.studentList.students.length == 0){
 
   this.endAddNewStudent = function(newStudent) {
     var className = self.studentList.className;
+    var classId = self.studentList._id;
     var newKid = newStudent;
     self.newStudent = {}
+    newKid.classId = classId;
     newKid.className = className;
-
-    classService.newStudentToAdd(newKid).then(function(response, err) {
+    var userId = self.user._id;
+    classService.newStudentToAdd(userId, newKid, classId).then(function(response, err) {
       newKid = response.data.students[(response.data.students.length - 1)];
       self.studentList.students.push(newKid);
       self.classInfo = self.studentList;
